create release 1.0.0.0
I've gone with 1.0.0.0 due to this being the first release in the open
source epoc.  Structurally, we could go with a patch level release, but
due to the change in the docker image build process as well as the
image's name and tag structure, I'm making this an epoch release.